chore(brochureware): drop changelog plugin and generate JSON via script
