+++ packages/brochureware/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "predev": "node scripts/collect-screenshots.js",
     "dev": "npm run build && vite",
-    "prebuild": "node scripts/collect-screenshots.js",
+    "prebuild": "node scripts/collect-screenshots.js && node scripts/generate-changelog.js",
     "build": "npm run prebuild && vite build",
     "build:dev": "vite build --mode development",
     "lint": "eslint .",
